update installation script and some tsserver stuff

+++ utils/install-lsp.sh
@@ -2,9 +2,42 @@
 
 SECONDS=0
 
+LOG_LEVEL=${LOG_LEVEL-"INFO"}
 source <(curl -s "https://gist.githubusercontent.com/cenk1cenk2/e03d8610534a9c78f755c1c1ed93a293/raw/3d61dc3718f3a3687d5990b9b5dc951198d29427/logger.sh")
 
 LSP_FOLDER=~/.config/nvim/lsp-servers
+
+HELP=$(
+	cat <<-END
+		Installs dependencies for language servers for neovim. Supports node, python, go, binary and vscode-extensions.
+
+		${0} [command]
+		help    shows this help
+		clean   deletes LSP_FOLDER at ${LSP_FOLDER} and performs a clean install
+
+		You can use LOG_LEVEL=DEBUG ${0} for debugging output.
+		You can use LOG_LEVEL=LIFETIME ${0} for less output.
+	END
+)
+
+log_this "[install-lsp]" "false" "lifetime" "bottom"
+
+if [[ -n $1 && $1 == "help" && -n $HELP ]]; then
+	echo "$HELP"
+	exit 0
+fi
+
+if [[ -n $1 && $1 == "clean" ]]; then
+	if [ -d "${LSP_FOLDER}" ]; then
+		log_warn "Removing LSP folder for clean install: ${LSP_FOLDER}"
+		rm -r "${LSP_FOLDER}"
+	else
+		log_error "LSP folder is not really a folder: ${LSP_FOLDER}"
+		ls -la "${LSP_FOLDER}"
+		exit 127
+	fi
+fi
+
 NPM_EXTENSIONS=(
 	"typescript;tsserver"
 	"typescript-language-server"
@@ -39,8 +72,6 @@ PYTHON_EXTENSIONS=(
 	"mypy"
 )
 
-log_this "[install-lsp]" "false" "lifetime" "bottom"
-
 log_info "LSP Folder: ${LSP_FOLDER}"
 mkdir -p ${LSP_FOLDER}
 cd ${LSP_FOLDER} || exit 127
@@ -65,11 +96,30 @@ function split_string() {
 
 function install_and_link_binaries() {
 	TYPE=$1
-	BASE_DIR=$2
-	INSTALL_COMMAND=$3
-	EXTENSIONS=($4)
+	EXTENSIONS=($2)
+
+	unset BASE_DIR
+	unset INSTALL_COMMAND
+	unset CLEAN_COMMAND
+	unset ALL_EXTENSIONS
+	unset ALL_BINARIES
 
 	log_start "Installing $TYPE packages..." "top"
+
+	if [ "$TYPE" = "node" ]; then
+		BASE_DIR="./node_modules/.bin"
+		INSTALL_COMMAND="yarn add --prod"
+	elif [ "$TYPE" = "python" ]; then
+		BASE_DIR="./venv/bin"
+		INSTALL_COMMAND="./venv/bin/pip3 install"
+	elif [ "$TYPE" = "go" ]; then
+		INSTALL_COMMAND="GOPATH=$(pwd) GOBIN=$(pwd) GO111MODULE=on go get -v"
+		CLEAN_COMMAND=("GOPATH=$(pwd) GO111MODULE=on go clean -modcache" "rm -rf src pkg 2>/dev/null")
+	else
+		log_error "Package type not supported: ${TYPE}"
+
+		exit 127
+	fi
 
 	ALL_EXTENSIONS=()
 	ALL_BINARIES=()
@@ -84,25 +134,40 @@ function install_and_link_binaries() {
 
 		eval "${INSTALL_COMMAND} ${ALL_EXTENSIONS[*]}"
 
-		log_info "Linking binaries: ${ALL_BINARIES[*]}"
-		for e in "${ALL_BINARIES[@]}"; do
-			if [[ -L "$LSP_FOLDER/$e" ]]; then
-				log_warn "Link already exists for $e, deleting it first."
-				rm "$LSP_FOLDER/${e}"
-			fi
+		if [ -n "${BASE_DIR}" ]; then
+			log_info "Linking binaries: ${ALL_BINARIES[*]}"
 
-			if [ "$e" != "false" ] && [ ! -f "${BASE_DIR}/${e}" ]; then
-				log_error "${e} is not a $TYPE binary."
-				log_info "Binaries listed as follows:"
-				ls -la "${BASE_DIR}"
-				exit 127
-			elif [ "$e" == "false" ]; then
-				log_warn "No binary for library."
-			else
-				log_info "Linking executable: ${BASE_DIR}/${e}"
-				ln -s "${BASE_DIR}/${e}" .
-			fi
-		done
+			for e in "${ALL_BINARIES[@]}"; do
+				if [[ -L "$LSP_FOLDER/$e" ]]; then
+					log_debug "Link already exists for $e, deleting it first."
+					rm "$LSP_FOLDER/${e}"
+				fi
+
+				if [ "$e" != "false" ] && [ ! -f "${BASE_DIR}/${e}" ]; then
+					log_error "${e} is not a $TYPE binary."
+					log_info "Binaries listed as follows:"
+					ls -la "${BASE_DIR}"
+					exit 127
+				elif [ "$e" == "false" ]; then
+					log_debug "No binary for library."
+				else
+					log_debug "Linking executable: ${BASE_DIR}/${e}"
+					ln -s "${BASE_DIR}/${e}" .
+				fi
+			done
+
+		fi
+
+		if [ -n "${CLEAN_COMMAND[*]}" ]; then
+			log_info "Executing post-installation cleaning: ${CLEAN_COMMAND[*]}"
+
+			for COMMAND in "${CLEAN_COMMAND[@]}"; do
+
+				eval "${COMMAND}"
+
+			done
+		fi
+
 	else
 		log_warn "No $TYPE extensions to install. Skipping..."
 	fi
@@ -123,7 +188,7 @@ function extract_archive() {
 	TMP_UNZIPPED_FOLDER=$2
 	COMPRESSION=$3
 
-	log_info "Unzipping file: $TMP_DOWNLOAD_PATH as $COMPRESSION"
+	log_debug "Unzipping file: $TMP_DOWNLOAD_PATH as $COMPRESSION"
 
 	if [ "${COMPRESSION}" == "tar_gz" ]; then
 		mkdir -p "${TMP_UNZIPPED_FOLDER}"
@@ -149,10 +214,15 @@ function download_binary() {
 	COMPRESSION=$2
 	BINARY=($3)
 
+	unset TMP_DOWNLOAD_PATH
+	unset TMP_UNZIPPED_FOLDER
+	unset ASSET_NAME
+	unset ASSET_TO_COPY
+
 	TMP_DOWNLOAD_PATH="/tmp/$(cat /dev/urandom | tr -cd 'a-f0-9' | head -c 32)"
 	TMP_UNZIPPED_FOLDER="/tmp/$(cat /dev/urandom | tr -cd 'a-f0-9' | head -c 32)"
 
-	log_start "Downloading binary: $BINARY from $URL" "top"
+	log_start "Downloading binary: ${BINARY[*]} from $URL" "top"
 
 	fetch_url "${URL}" "${TMP_DOWNLOAD_PATH}"
 
@@ -165,10 +235,10 @@ function download_binary() {
 		cp "${ASSET_TO_COPY}" "${LSP_FOLDER}"
 
 		ASSET_NAME=$(basename -- "$ASSET_TO_COPY")
-		log_info "chmod +x: ${ASSET_NAME}"
+		log_debug "chmod +x: ${ASSET_NAME}"
 		chmod +x "${LSP_FOLDER}/${ASSET_NAME}"
 
-		log_info "Asset copied to lsp folder: ${ASSET_NAME}"
+		log_info "Binary copied to lsp folder: ${ASSET_NAME}"
 	done
 
 	rm "${TMP_UNZIPPED_FOLDER}" -r
@@ -183,6 +253,9 @@ function download_extension() {
 	SUBPATH=$4
 	CHMOD=(${5})
 
+	unset TMP_DOWNLOAD_PATH
+	unset TMP_UNZIPPED_FOLDER
+
 	TMP_DOWNLOAD_PATH="/tmp/$(cat /dev/urandom | tr -cd 'a-f0-9' | head -c 32)"
 	TMP_UNZIPPED_FOLDER="/tmp/$(cat /dev/urandom | tr -cd 'a-f0-9' | head -c 32)"
 
@@ -193,7 +266,7 @@ function download_extension() {
 	extract_archive "${TMP_DOWNLOAD_PATH}" "${TMP_UNZIPPED_FOLDER}" "${COMPRESSION}"
 
 	if [ -d "${LSP_FOLDER}/${EXTENSION}" ] || [ -f "${LSP_FOLDER}/${EXTENSION}" ]; then
-		log_warn "Root for extension already found, removing: ${EXTENSION}"
+		log_debug "Root for extension already found, removing: ${EXTENSION}"
 		rm "${LSP_FOLDER}/${EXTENSION}" -r
 	fi
 
@@ -202,48 +275,46 @@ function download_extension() {
 	mv "${TMP_UNZIPPED_FOLDER}/${SUBPATH:-'.'}/" "${LSP_FOLDER}/${EXTENSION}"
 
 	for e in "${CHMOD[@]}"; do
-		log_info "chmod +x -R: ${EXTENSION}/${e}"
+		log_debug "chmod +x -R: ${EXTENSION}/${e}"
 		chmod +x -R "${LSP_FOLDER}/${EXTENSION}/${e}"
 	done
 }
 
 # for npm based extensions
-install_and_link_binaries "NPM" "./node_modules/.bin" "yarn add --prod" "${NPM_EXTENSIONS[*]}"
+install_and_link_binaries "node" "${NPM_EXTENSIONS[*]}"
 
 # for go based extensions
-for e in "${GO_EXTENSIONS[@]}"; do
-	log_start "Installing GO binary: ${e}" "top"
-	split_string ${e}
-
-	GOPATH=$(pwd) GOBIN=$(pwd) GO111MODULE=on go get -v "${PACKAGE_NAME}"
-	GOPATH=$(pwd) GO111MODULE=on go clean -modcache
-	rm -rf src pkg 2>/dev/null
-
-	log_finish "Installed GO binary: ${e}"
-done
+install_and_link_binaries "go" "${GO_EXTENSIONS[*]}"
 
 # for python based extensions
+log_start "Initiating new python environment." "top"
 python3 -m venv ./venv
 ./venv/bin/pip3 install -U pip
+log_finish "Now in new python environment."
 
-install_and_link_binaries "Python" "./venv/bin" "./venv/bin/pip3 install" "${PYTHON_EXTENSIONS[*]}"
+install_and_link_binaries "python" "${PYTHON_EXTENSIONS[*]}"
 
 ## install custom stuff with curl, for compiled binaries mostly
 
-log_start "Installing custom assets with curl..."
+log_start "Installing custom assets and vscode-extensions..." "top"
 
 # install shell-check here
-VERSION=v0.7.1
+VERSION="v0.7.1"
 download_binary "https://github.com/koalaman/shellcheck/releases/download/${VERSION}/shellcheck-${VERSION}.linux.x86_64.tar.xz" "tar_xz" "shellcheck-${VERSION}/shellcheck"
 
+# install lua language server
 VERSION="1.18.1"
 download_extension "https://github.com/sumneko/vscode-lua/releases/download/v$VERSION/lua-$VERSION.vsix" "zip" "lua-language-server" "extension/server" "bin"
 
+# install eslint language server, dont upgrade this until figuring out how to accept the prompt
 VERSION="2.1.0"
 download_extension "https://github.com/microsoft/vscode-eslint/releases/download/release%2F$VERSION-next.1/vscode-eslint-$VERSION.vsix" "zip" "eslint-language-server" "extension/server/out"
 
+# install tailwinds language server
 VERSION="0.5.9"
 download_extension "https://github.com/tailwindlabs/tailwindcss-intellisense/releases/download/v$VERSION/vscode-tailwindcss-$VERSION.vsix" "zip" "tailwindcss-language-server" "extension/dist/server"
 
-log_finish "Installed lsp dependencies." "top"
-log_info "$((SECONDS / 60)) minutes and $((SECONDS % 60)) seconds elapsed."
+log_finish "Installed custom assets and vscode-extensions."
+
+# end
+log_finish "Installed lsp dependencies in $((SECONDS / 60)) minutes and $((SECONDS % 60)) seconds." "top"
